Update paths and API functions inline with Electron changes
Because of some major changes in the Electron world, `path` can no longer be accessed in preload scripts, so I must cycle between main and renderer.
This also optimizes the grid resizing + makes it handle the new `async`/`await` flow because of the API changes. The same is applied to the random audio selector and `MainWindow`.